@@ -2,7 +2,6 @@
   "name": "fsh-online",
   "version": "1.0.0",
   "private": true,
-  "homepage": "https://fshschool.org/FSHOnline/",
   "license": "Apache-2.0",
   "dependencies": {
     "@material-ui/core": "^4.10.2",
@@ -19,52 +18,40 @@
     "gofsh": "^2.3.1",
     "jszip": "^3.10.1",
     "lodash": "^4.17.21",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-codemirror2": "^7.3.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "tar-stream": "^2.1.3"
   },
   "devDependencies": {
-    "cors": "^2.8.5",
+    "@babel/eslint-parser": "^7.24.5",
+    "@babel/preset-react": "^7.24.1",
+    "@vitejs/plugin-react": "^4.2.1",
     "cross-env": "^7.0.3",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-prettier": "^3.1.3",
-    "express": "^4.18.2",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "fake-indexeddb": "^3.1.1",
     "nock": "^13.4.0",
     "null-loader": "^4.0.1",
     "prettier": "^2.0.5",
-    "react-app-rewired": "^2.1.6",
-    "react-scripts": "^4.0.3",
-    "serve-static": "^1.15.0"
-  },
-  "overrides": {
-    "antlr4": "4.13.1-patch-1",
-    "fsh-sushi": {
-      "https-proxy-agent": "5.0.0",
-      "ini": "4.0.0"
-    },
-    "fhir-package-loader": {
-      "https-proxy-agent": "5.0.0"
-    },
-    "react-dev-utils": {
-      "react-error-overlay": "6.0.9"
-    }
+    "vite": "^5.2.10",
+    "vite-plugin-node-polyfills": "^0.21.0"
   },
   "scripts": {
-    "start": "cross-env NODE_OPTIONS=--openssl-legacy-provider react-app-rewired start",
-    "build": "cross-env NODE_OPTIONS=--openssl-legacy-provider react-app-rewired build",
-    "serve-build": "node scripts/serve-build.js",
-    "lint": "eslint \"./src/**/*.js\"",
-    "lint-fix": "eslint \"./src/**/*.js\" --fix",
-    "prettier": "prettier --check \"**/*.{js,ts}\"",
-    "test": "react-app-rewired test",
-    "eject": "react-app-rewired eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint \"./src/**/*.{js,jsx}\"",
+    "lint-fix": "eslint \"./src/**/*.{js,jsx}\" --fix",
+    "prettier": "prettier --check \"**/*.{js,ts,jsx,tsx}\"",
+    "test": "jest"
   },
   "jest": {
     "moduleNameMapper": {
