@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import App from './App';
 
-export default function AppRouter(props) {
+export default function AppRouter() {
   return (
     <Switch>
       <Route path="/share/:text" component={App} />
